Removed more lines
